@@ -1,6 +1,4 @@
-
 function Home() {
-    console.log('Home page')
     return (
         <div>
             <h2>Home</h2>
@@ -9,7 +7,6 @@ function Home() {
     )
 }
 function About() {
-    console.log('About page');
     return (
         <div>
             <h2>About</h2>
@@ -18,11 +15,10 @@ function About() {
     )
 }
 function PageNotFound() {
-    console.log('This page opens when we don\'t have any matching page');
     return (
         <>
             <h2>OOPs Page not found...</h2>
-            <p> No page whith this url is present</p>
+            <p>No page with this url is present</p>
         </>
     )
 }
